Remove the old object menu
