+++ sensemap/src/components/ObjectMenu/index.tsx
@@ -1,33 +1,42 @@
 import * as React from 'react';
 import { connect } from 'react-redux';
-import { Menu } from 'semantic-ui-react';
+import { State, actions, ActionProps, mapDispatch } from '../../types';
+import { Menu, Popup } from 'semantic-ui-react';
+import Card from '../SVGIcon/Card';
+import Box from '../SVGIcon/Box';
+import Unbox from '../SVGIcon/Unbox';
+import Inspector from '../SVGIcon/Inspector';
+import Edge from '../SVGIcon/Edge';
+import RemoveEdge from '../SVGIcon/RemoveEdge';
+import Copy from '../SVGIcon/Copy';
+import Delete from '../SVGIcon/Delete';
+import BatchTag from '../SVGIcon/BatchTag';
+import BoxCard from '../SVGIcon/BoxCard';
+import OpenBox from '../SVGIcon/OpenBox';
+import Eject from '../SVGIcon/Eject';
 import * as T from '../../types';
-import { actions, ActionProps, mapDispatch } from '../../types';
 import * as SL from '../../types/selection';
-import * as CS from '../../types/cached-storage';
+import * as SM from '../../types/sense-map';
+import * as SO from '../../types/sense-object';
 import * as OE from '../../types/object-editor';
+import * as CS from '../../types/cached-storage';
+import * as F from '../../types/sense/focus';
+// TODO: use UUID v4
+import * as U from '../../types/utils';
 import { cardData } from '../../types/sense/card';
 import { boxData } from '../../types/sense/box';
-import * as F from '../../types/sense/focus';
-import * as I from '../../types/input';
-// TODO: use UUID v4
-import { objectId } from '../../types/utils';
+import './index.css';
 
 interface StateFromProps {
-  selection:   SL.State;
-  senseObject: T.State['senseObject'];
-  scope:       T.State['senseMap']['scope'];
-  senseMap:    T.State['senseMap'];
-  input:       T.State['input'];
-  editor:      T.State['editor'];
+  selection: SL.State;
+  senseObject: SO.State;
+  senseMap: SM.State;
+  editor: OE.State;
 }
 
-interface OwnProps {
-  // tslint:disable-next-line:no-any
-  style: any;
-}
+interface OwnProps {}
 
-type Props = StateFromProps & ActionProps & OwnProps;
+type Props = StateFromProps & OwnProps & ActionProps;
 
 const selectedCardsAndBoxes:
   (props: Props) => { cards: T.ObjectID[], boxes: T.ObjectID[] } =
@@ -49,27 +58,34 @@ const selectedCardsAndBoxes:
   );
 
 class ObjectMenu extends React.PureComponent<Props> {
-  constructor(props: Props) {
-    super(props);
-    this.canAddCard = this.canAddCard.bind(this);
-    this.handleAddCard = this.handleAddCard.bind(this);
-    this.canRemoveCard = this.canRemoveCard.bind(this);
-    this.handleRemoveCard = this.handleRemoveCard.bind(this);
-    this.handleUnbox = this.handleUnbox.bind(this);
+  handleCreateCard(): void {
+    const { actions: acts } = this.props;
+    const data = cardData({ id: U.objectId() });
+    acts.senseObject.updateCard(data);
+    acts.editor.focusObject(F.focusCard(data.id));
+    acts.editor.changeStatus(OE.StatusType.SHOW);
   }
 
-  canCreateBox(): Boolean {
-    return this.props.scope.type === T.MapScopeType.FULL_MAP;
+  canCreateBox(): boolean {
+    return this.props.senseMap.scope.type === T.MapScopeType.FULL_MAP;
+  }
+
+  handleBox(): void {
+    const { actions: acts } = this.props;
+    const data = boxData({ id: U.objectId() });
+    acts.senseObject.updateBox(data);
+    acts.editor.focusObject(F.focusBox(data.id));
+    acts.editor.changeStatus(OE.StatusType.SHOW);
   }
 
   canUnbox(): Boolean {
-    return this.props.scope.type === T.MapScopeType.BOX;
+    return this.props.senseMap.scope.type === T.MapScopeType.BOX;
   }
 
   handleUnbox(): void {
-    switch (this.props.scope.type) {
+    switch (this.props.senseMap.scope.type) {
       case T.MapScopeType.BOX: {
-        const box = this.props.scope.box;
+        const box = this.props.senseMap.scope.box;
         if (!box) {
           throw Error('Scope BOX without a box ID.');
         }
@@ -81,51 +97,11 @@ class ObjectMenu extends React.PureComponent<Props> {
     }
   }
 
-  canAddCard(): Boolean {
-    if (this.props.scope.type !== T.MapScopeType.FULL_MAP) {
-      return false;
-    }
-    const { cards, boxes } = selectedCardsAndBoxes(this.props);
-    return cards.length >= 1 && boxes.length === 1;
-  }
-
-  handleAddCard(): void {
-    if (!this.canAddCard()) {
-      return;
-    }
-    const { cards, boxes } = selectedCardsAndBoxes(this.props);
-    const object = CS.getObject(this.props.senseObject, boxes[0]);
-    this.props.actions.senseObject.addCardsToBox(cards, object.data);
-    return;
-  }
-
-  canRemoveCard(): Boolean {
-    return this.props.scope.type === T.MapScopeType.BOX
-      && this.props.selection.length >= 1;
-  }
-
-  canDeleteCard(): Boolean {
-    const { cards } = selectedCardsAndBoxes(this.props);
-    return cards.length === 1;
-  }
-
-  handleRemoveCard(): void {
-    if (!this.canRemoveCard()) {
-      return;
-    }
-    switch (this.props.scope.type) {
-      case T.MapScopeType.BOX: {
-        const cards = this.props.selection;
-        const box   = this.props.scope.box;
-        if (!box) {
-          throw Error('This cannot happen: map scope has type BOX with null box ID.');
-        }
-        this.props.actions.senseObject.removeCardsFromBox(cards, box);
-        break;
-      }
-      case T.MapScopeType.FULL_MAP:
-      default:
-    }
+  findEdgeID(from: T.ObjectID, to: T.ObjectID): T.EdgeID[] | null {
+    const r = Object.values(CS.toStorage(this.props.senseObject).edges)
+      .filter(edge => (edge.from === from && edge.to === to) || (edge.from === to && edge.to === from))
+      .map(edge => edge.id);
+    return r.length === 0 ? null : r;
   }
 
   canCreateEdge(): boolean {
@@ -141,13 +117,6 @@ class ObjectMenu extends React.PureComponent<Props> {
     const from = this.props.selection[0];
     const to   = this.props.selection[1];
     this.props.actions.senseObject.createEdge(map, from, to);
-  }
-
-  findEdgeID(from: T.ObjectID, to: T.ObjectID): T.EdgeID[] | null {
-    const r = Object.values(CS.toStorage(this.props.senseObject).edges)
-      .filter(edge => (edge.from === from && edge.to === to) || (edge.from === to && edge.to === from))
-      .map(edge => edge.id);
-    return r.length === 0 ? null : r;
   }
 
   canRemoveEdge(): boolean {
@@ -168,132 +137,240 @@ class ObjectMenu extends React.PureComponent<Props> {
     r.forEach(edge => this.props.actions.senseObject.removeEdge(map, edge));
   }
 
+  canCopy(): boolean {
+    const { cards, boxes } = selectedCardsAndBoxes(this.props);
+    return cards.length + boxes.length === 1;
+  }
+
+  canDeleteCard(): boolean {
+    const { cards } = selectedCardsAndBoxes(this.props);
+    return cards.length === 1;
+  }
+
+  async handleDeleteCard() {
+    const { actions: acts, senseObject, selection } = this.props;
+    const { id, data } = CS.getObject(senseObject, selection[0]);
+    const card = CS.getCard(senseObject, data);
+    // remove the card container object
+    await acts.senseObject.removeObject(id);
+    // clear any local modifications
+    acts.cachedStorage.removeCard(card);
+    acts.editor.focusObject(F.focusNothing());
+  }
+
+  canBatchTag(): boolean {
+    const { cards, boxes } = selectedCardsAndBoxes(this.props);
+    return cards.length + boxes.length > 1;
+  }
+
+  canAddCard(): boolean {
+    if (this.props.senseMap.scope.type !== T.MapScopeType.FULL_MAP) {
+      return false;
+    }
+    const { cards, boxes } = selectedCardsAndBoxes(this.props);
+    return cards.length >= 1 && boxes.length === 1;
+  }
+
+  handleAddCard(): void {
+    if (!this.canAddCard()) {
+      return;
+    }
+    const { cards, boxes } = selectedCardsAndBoxes(this.props);
+    const object = CS.getObject(this.props.senseObject, boxes[0]);
+    this.props.actions.senseObject.addCardsToBox(cards, object.data);
+    return;
+  }
+
+  canOpenBox(): boolean {
+    const { boxes } = selectedCardsAndBoxes(this.props);
+    return boxes.length === 1;
+  }
+
+  handleOpenBox(): void {
+    const { actions: acts, selection, senseObject } = this.props;
+    const objectId = selection[0];
+    const obj = CS.getObject(senseObject, objectId);
+    acts.selection.clearSelection();
+    acts.senseMap.openBox(obj.data);
+  }
+
+  canRemoveCard(): Boolean {
+    return this.props.senseMap.scope.type === T.MapScopeType.BOX
+      && this.props.selection.length >= 1;
+  }
+
+  handleRemoveCard(): void {
+    if (!this.canRemoveCard()) {
+      return;
+    }
+    switch (this.props.senseMap.scope.type) {
+      case T.MapScopeType.BOX: {
+        const cards = this.props.selection;
+        const box   = this.props.senseMap.scope.box;
+        if (!box) {
+          throw Error('This cannot happen: map scope has type BOX with null box ID.');
+        }
+        this.props.actions.senseObject.removeCardsFromBox(cards, box);
+        break;
+      }
+      case T.MapScopeType.FULL_MAP:
+      default:
+    }
+  }
+
   render() {
-    const { style, actions: acts, senseObject, selection, input, editor } = this.props;
-    const isMultiSelectable = I.isMultiSelectable(input);
+    const { actions: acts, editor } = this.props;
+    const canCreateBox = this.canCreateBox();
+    const canUnbox = this.canUnbox();
+    const isInspectorOpen = editor.status === OE.StatusType.SHOW;
+    const canCreateEdge = this.canCreateEdge();
+    const canRemoveEdge = this.canRemoveEdge();
+    const canCopy = this.canCopy();
+    const canDeleteCard = this.canDeleteCard();
+    const canBatchTag = this.canBatchTag();
+    const canAddCard = this.canAddCard();
+    const canOpenBox = this.canOpenBox();
+    const canRemoveCard = this.canRemoveCard();
+    const showSecondMenu =
+      canCopy || canDeleteCard || canBatchTag || canAddCard || canOpenBox || canRemoveCard;
+    // setup the popup props and force the position to be a string literal
+    const popupProps = { inverted: true, position: 'bottom center' as 'bottom center' };
 
     return (
-      <Menu vertical style={style}>
-        <Menu.Item>{
-          selection.length === 0
-            ? '功能選單'
-            : selection.length === 1
-              ? isMultiSelectable
-                ? '多選模式'
-                : '按 Shift/Ctrl 選取多張卡片'
-              : `選取了 ${selection.length} 張卡片`
-        }</Menu.Item>
-        {
-          this.canCreateBox() &&
-          <Menu.Item
-            name="create-box"
-            onClick={() => {
-              const data = boxData({ id: objectId() });
-              acts.senseObject.updateBox(data);
-              acts.editor.focusObject(F.focusBox(data.id));
-              acts.editor.changeStatus(OE.StatusType.SHOW);
-            }}
-          >
-            新增 Box
-          </Menu.Item>
-        }
-        <Menu.Item
-          name="create-card"
-          onClick={() => {
-            const data = cardData({ id: objectId() });
-            acts.senseObject.updateCard(data);
-            acts.editor.focusObject(F.focusCard(data.id));
-            acts.editor.changeStatus(OE.StatusType.SHOW);
-          }}
-        >
-          新增卡片
-        </Menu.Item>
-        <Menu.Item
-          name="edit"
-          onClick={() => {
-            if (editor.status === OE.StatusType.HIDE) {
-              acts.editor.changeStatus(OE.StatusType.SHOW);
-            } else {
-              acts.editor.changeStatus(OE.StatusType.HIDE);
+      <div className="sense-object-menu">
+        <Menu compact inverted icon>
+          <Popup
+            {...popupProps}
+            trigger={<Menu.Item onClick={() => this.handleCreateCard()}><Card /></Menu.Item>}
+            content="New Card"
+          />
+          <Popup
+            {...popupProps}
+            trigger={
+              <Menu.Item
+                disabled={!canUnbox && !canCreateBox}
+                onClick={() =>
+                  canUnbox
+                    ? this.handleUnbox()
+                    : this.handleBox()
+                }
+              >
+                {canUnbox ? <Unbox /> : <Box />}
+              </Menu.Item>
             }
-          }}
-        >
-          編輯器
-        </Menu.Item>
+            content={
+              canUnbox
+                ? 'Unbox'
+                : canCreateBox
+                  ? 'New Box'
+                  : 'Can\'t Create a Box in a Box'
+            }
+          />
+          <Popup
+            {...popupProps}
+            trigger={
+              <Menu.Item
+                active={isInspectorOpen}
+                onClick={() =>
+                  acts.editor.changeStatus(
+                    isInspectorOpen
+                      ? OE.StatusType.HIDE
+                      : OE.StatusType.SHOW
+                  )
+                }
+              >
+                <Inspector />
+              </Menu.Item>
+            }
+            content="Edit"
+          />
+          <Popup
+            {...popupProps}
+            trigger={
+              <Menu.Item
+                disabled={!canRemoveEdge && !canCreateEdge}
+                onClick={() =>
+                  canRemoveEdge
+                    ? this.handleRemoveEdge()
+                    : this.handleCreateEdge()
+                }
+              >
+                {canRemoveEdge ? <RemoveEdge /> : <Edge />}
+              </Menu.Item>
+            }
+            content={
+              canRemoveEdge
+                ? 'Remove Edge'
+                : canCreateEdge
+                  ? 'Add Edge'
+                  : 'Select Two Objects First'
+            }
+          />
+        </Menu>
         {
-          this.canAddCard() &&
-          <Menu.Item
-            name="addCard"
-            onClick={this.handleAddCard}
-          >
-            加入
-          </Menu.Item>
+          showSecondMenu &&
+          <Menu compact inverted icon>
+            {
+              // the copy action is not implemented
+              canCopy &&
+              <Popup
+                {...popupProps}
+                trigger={<Menu.Item disabled onClick={U.noop}><Copy /></Menu.Item>}
+                content="Copy"
+              />
+            }
+            {
+              canDeleteCard &&
+              <Popup
+                {...popupProps}
+                trigger={<Menu.Item onClick={() => this.handleDeleteCard()}><Delete /></Menu.Item>}
+                content="Delete"
+              />
+            }
+            {
+              canBatchTag &&
+              <Popup
+                {...popupProps}
+                trigger={<Menu.Item disabled onClick={U.noop}><BatchTag /></Menu.Item>}
+                content="Batch Tag"
+              />
+            }
+            {
+              canAddCard &&
+              <Popup
+                {...popupProps}
+                trigger={<Menu.Item onClick={() => this.handleAddCard()}><BoxCard /></Menu.Item>}
+                content="Box Cards"
+              />
+            }
+            {
+              canOpenBox &&
+              <Popup
+                {...popupProps}
+                trigger={<Menu.Item onClick={() => this.handleOpenBox()}><OpenBox /></Menu.Item>}
+                content="Open Box"
+              />
+            }
+            {
+              canRemoveCard &&
+              <Popup
+                {...popupProps}
+                trigger={<Menu.Item onClick={() => this.handleRemoveCard()}><Eject /></Menu.Item>}
+                content="Eject"
+              />
+            }
+          </Menu>
         }
-        {
-          this.canRemoveCard() &&
-          <Menu.Item
-            name="removeCard"
-            onClick={this.handleRemoveCard}
-          >
-            退出
-          </Menu.Item>
-        }
-        {
-          this.canDeleteCard() &&
-          <Menu.Item
-            name="deleteCard"
-            onClick={async () => {
-              const { id, data } = CS.getObject(senseObject, selection[0]);
-              const card = CS.getCard(senseObject, data);
-              // remove the card container object
-              await acts.senseObject.removeObject(id);
-              // clear any local modifications
-              acts.cachedStorage.removeCard(card);
-              acts.editor.focusObject(F.focusNothing());
-            }}
-          >
-            刪除
-          </Menu.Item>
-        }
-        {
-          this.canUnbox() &&
-          <Menu.Item
-            name="unbox"
-            onClick={this.handleUnbox}
-          >
-            Unbox
-          </Menu.Item>
-        }
-        {
-          this.canCreateEdge() &&
-          <Menu.Item
-            name="createEdge"
-            onClick={() => this.handleCreateEdge()}
-          >
-            連線
-          </Menu.Item>
-        }
-        {
-          this.canRemoveEdge() &&
-          <Menu.Item
-            name="deleteEdge"
-            onClick={() => this.handleRemoveEdge()}
-          >
-            刪除連線
-          </Menu.Item>
-        }
-      </Menu>
+      </div>
     );
   }
 }
 
 export default connect<StateFromProps, ActionProps>(
-  (state: T.State) => ({
+  (state: State) => ({
     selection: state.selection,
-    scope: state.senseMap.scope,
     senseObject: state.senseObject,
     senseMap: state.senseMap,
-    input: state.input,
     editor: state.editor,
   }),
   mapDispatch({ actions }),
