--- sensemap/src/components/ObjectMenu2/index.css
@@ -1,28 +1,0 @@
-.object-menu-2 > .ui.inverted.menu {
-  box-shadow: 0 1px 4px 0 rgba(0, 0, 0, 0.5);
-}
-
-.object-menu-2 > .ui.inverted.menu:first-child {
-  margin-right: 2px;
-}
-
-.object-menu-2 > .ui.inverted.menu .item {
-  padding: 0.428571em;
-}
-
-.object-menu-2 > .ui.inverted.menu .item.disabled {
-  color: rgba(225,225,225,.3) !important;
-}
-
-.object-menu-2 > .ui.inverted.menu .item:before {
-  width: 0;
-}
-
-.object-menu-2 > .ui.inverted.menu .item > .sense-svg-icon {
-  width: 1.714285em;
-  height: 1.714285em;
-}
-
-.ui.inverted.popup {
-  padding: 0.428571em 0.714285em;
-}
